+++ _admin.php
@@ -1,16 +1,15 @@
 <?php
 /**
  * @brief improve, a plugin for Dotclear 2
- * 
+ *
  * @package Dotclear
  * @subpackage Plugin
- * 
+ *
  * @author Jean-Christian Denis and contributors
- * 
+ *
  * @copyright Jean-Christian Denis
  * @copyright GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html
  */
-
 $core->blog->settings->addNamespace('improve');
 
 $core->addBehavior('adminDashboardFavorites', ['ImproveBehaviors', 'adminDashboardFavorites']);
@@ -22,6 +21,7 @@ $core->addBehavior('improveAddAction', ['ImproveActionGitshields', 'create']);
 $core->addBehavior('improveAddAction', ['ImproveActionLicensefile', 'create']);
 //$core->addBehavior('improveAddAction', ['ImproveActionLicense', 'create']);
 $core->addBehavior('improveAddAction', ['ImproveActionNewline', 'create']);
+$core->addBehavior('improveAddAction', ['ImproveActionPhpcsfixer', 'create']);
 $core->addBehavior('improveAddAction', ['ImproveActionPhpheader', 'create']);
 $core->addBehavior('improveAddAction', ['ImproveActionTab', 'create']);
 $core->addBehavior('improveAddAction', ['ImproveActionZip', 'create']);